[zone/storage] Break down `start_switch()`
